change display architecture, statistics architecture, change dataframe functions
