+++ validations/person_dataset_validator.py
@@ -1,0 +1,9 @@
+from pandas import DataFrame
+
+
+def validate_person_dataset(dataset: DataFrame):
+    copied_dataset = dataset.copy()
+
+
+def validate_repeatable_words_in_dataset_by_column(dataset: DataFrame, column: str):
+    pass
